+++ test/polisher_test.cpp
@@ -11,7 +11,7 @@
 #include "edlib.h"  // NOLINT
 #include "gtest/gtest.h"
 
-std::atomic<std::uint64_t> biosoup::Sequence::num_objects{0};
+std::atomic<std::uint32_t> biosoup::Sequence::num_objects{0};
 
 namespace racon {
 namespace test {
